# atlasio-monorepo/apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/courses/{courseId}/lessons": {
      "post": {
        "operationId": "LessonsController_create",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "lessons"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "get": {
        "operationId": "LessonsController_list",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lessons"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/courses/{courseId}/lessons/{lessonId}": {
      "get": {
        "operationId": "LessonsController_get",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lessons"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/courses/{courseId}/lessons/{lessonId}/complete": {
      "post": {
        "operationId": "LessonsController_complete",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lessons"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/courses/{courseId}/progress": {
      "get": {
        "operationId": "LessonsController_progress",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lessons"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/me/progress/bulk": {
      "post": {
        "operationId": "LessonsController_bulkProgress",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lessons"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Atlasio API",
    "description": "Atlasio Kurumsal Uzaktan Eğitim API",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "tenant": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID"
      }
    },
    "schemas": {
...
      "CreateLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Giriş"
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "content": {
            "type": "string",
            "example": "Hoş geldin!"
          },
          "isPreview": {
            "type": "boolean",
            "example": true
          },
          "videoUrl": {
            "type": "string",
            "example": "https://example.com/video.mp4"
          }
        },
        "required": [
          "title"
        ]
      },
...
    }
  }
}